% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rssistat-package.R
\docType{package}
\name{rssistat-package}
\alias{rssistat}
\alias{rssistat-package}
\title{rssistat: ISTAT RSS Feed Discovery and Analysis}
\description{
Discover and analyze RSS feeds from ISTAT (Italian National Institute of Statistics) website. Provides tools for scanning web pages to find RSS/Atom feed URLs, parsing feed content, and summarizing news.
}
\author{
\strong{Maintainer}: Giampaolo Montaletti \email{giampaolo.montaletti@gmail.com}

}
\keyword{internal}
